requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
Pillow
tqdm
streamlit
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
requests